move helpers to utils directory
